Added support for locking and unlocking dialogs to disable user input when needed
Added locking icon style and various callbacks to track the behaviour of the dialog while it is being locked or unlocked.
Added control over the cursor on buttons depending on whether the buttons can actually be used
Updated samples to include a locking dialog example, plus included the locked icon in the assets folder.